requirements.txt
pandas
numpy
scikit-learn
xgboost
mlflow
fastapi
uvicorn
python-dotenv
matplotlib
seaborn
pyyaml
joblib
python-multipart